code/client/requirements.txt
pyaudio==0.2.14
opencv-python
torchaudio
requests==2.31.0
pywin32
psutil
keyboard==0.13.5
comtypes
pycaw
ffpyplayer
PyQt5
#PyQtWebEngine
pyautogui
pandas
pypiwin32==223
pillow==10.3.0